+++ app/components/ApiList.vue
@@ -30,7 +30,7 @@ const keyList = ref([]);
 const createKey = ref(false);
 
 const { data, error } = await useFetch("/api/auth/api-key/list");
-keyList.value = data.value;
+keyList.value = data.value.apiKeys;
 
 const handleKeyDeleted = async (id) => {
     const keyToDelete = keyList.value.find((key) => key.id === id);